Changes in DAA - Changed class 'Item' variable weight from float to int - Changed inputs from float to int accordingly in A3 and A4 - Optimized A4 - Removed branch and bound from A4

+++ LP3/DAA/src/A3.java
@@ -51,7 +51,7 @@ public class A3 {
             System.out.print("Enter profit for item " + (i+1) + ": ");
             int val = sc.nextInt();
             System.out.print("Enter weight for item " + (i+1) + ": ");
-            float w = sc.nextFloat();
+            int w = sc.nextInt();
             items[i] = new Item(val, w);
         }
